Trait diesel_async::UpdateAndFetchResults

pub trait UpdateAndFetchResults<Changes, Output>: AsyncConnectionwhere
    Changes: Identifiable + HasTable,{
    // Required method
    fn update_and_fetch<'life0, 'async_trait>(
        &'life0 mut self,
        changeset: Changes
    ) -> Pin<Box<dyn Future<Output = QueryResult<Output>> + Send + 'async_trait>>
       where Changes: 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
}

A trait defining how to update a record and fetch the updated entry on a certain backend.

The only case where it is required to work with this trait is while implementing a new connection type. Otherwise use SaveChangesDsl

For implementing this trait for a custom backend:

The Changes generic parameter represents the changeset that should be stored

The Output generic parameter represents the type of the response.

Required Methods

fn update_and_fetch<'life0, 'async_trait>( &'life0 mut self, changeset: Changes ) -> Pin<Box<dyn Future<Output = QueryResult<Output>> + Send + 'async_trait>>where Changes: 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

See the traits documentation.