Enum diesel_async::pooled_connection::RecyclingMethod

pub enum RecyclingMethod<C> {
    Fast,
    Verified,
    CustomQuery(Cow<'static, str>),
    CustomFunction(Box<RecycleCheckCallback<C>>),
}

Possible methods of how a connection is recycled.

Variants

Fast

Only check for open transactions when recycling existing connections Unless you have special needs this is a safe choice.

If the database connection is closed you will recieve an error on the first place you actually try to use the connection

Verified

In addition to checking for open transactions a test query is executed

This is slower, but guarantees that the database connection is ready to be used.

CustomQuery(Cow<'static, str>)

Like Verified but with a custom query

CustomFunction(Box<RecycleCheckCallback<C>>)

Like Verified but with a custom callback that allows to perform more checks

The connection is only recycled if the callback returns Ok(())